Legacy iWork '09 tables: share the extractor; Pages reads tables too
Lifts the '09 `<sf:tabular-model>` table extractor into SwiftTextIWA
(IWALegacyTableReader), along with TSTTable.trimmedToUsedRange(), so the
Numbers and Pages legacy readers share one decoder — the '09 table schema
is identical across iWork apps.

- PagesLegacyParser now extracts legacy tables too (previously body text
  only), appending each (cropped to its used range) as a table-bearing
  paragraph after the body, so `swifttext pages` surfaces tables from '09
  .pages files.
- NumbersParser's legacy branch uses the shared reader; the now-redundant
  NumbersLegacyParser and its unused error case are removed.

Verified: a legacy .pages with an <sf:tabular-model> renders its table to
Markdown; the real '09 .numbers glossary still reads unchanged. 443 tests pass.
